Conclude annotations
